pyproject.toml
[project]
name = "legal-workflow-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "sentence-transformers>=3.0.0",
    "pgvector>=0.3.0",
    "rank-bm25>=0.2.2",
]